backend/requirements.txt
pytz==2021.1
sqlparse==0.4.2
asgiref==3.4.1
pillow==8.3.2
djoser
djangorestframework_simplejwt
psycopg2-binary==2.9.1
gunicorn==20.0.4
django-filter